+++ Therapy/Data/DataModel/TasksResponse.swift
@@ -20,18 +20,18 @@ struct TasksResponse: Codable {
     }
     
     struct Task: Codable {
-        let breathCount: Int?
+        let breathCount: Float?
         let moodChoices: [String]?
         let dueDate: String?
         
         init(from decoder: any Decoder) throws {
             let container = try decoder.container(keyedBy: CodingKeys.self)
-            self.breathCount = try container.decode(Int.self, forKey: .breathCount)
+            self.breathCount = try container.decode(Float.self, forKey: .breathCount)
             self.moodChoices = try container.decode([String].self, forKey: .moodChoices)
             self.dueDate = try container.decode(String.self, forKey: .dueDate)
         }
         
-        init(breathCount: Int? = nil, moodChoices: [String]? = nil, dueDate: String? = nil) {
+        init(breathCount: Float? = nil, moodChoices: [String]? = nil, dueDate: String? = nil) {
             self.breathCount = breathCount
             self.moodChoices = moodChoices
             self.dueDate = dueDate
